# apps/js/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "BeliMudah API",
		"version": "1.0",
		"description": "E-commerce API for the BeliMudah storefront.",
		"contact": {
			"name": "Ryan Suhartanto",
			"url": "https://github.com/ryansuhartanto/koda-b8-backend",
			"email": "suhartanto@kekkon.nexus"
		},
		"license": {
			"name": "MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3002"
		}
	],
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		},
		"schemas": {
			"Problem": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer"
					},
					"detail": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"status"
				]
			},
...
			"Address": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"label": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"postal_code": {
						"type": "string"
					},
					"is_default": {
						"type": "boolean"
					}
				},
				"required": [
					"address",
					"city",
					"id",
					"is_default",
					"label",
					"name",
					"phone",
					"postal_code",
					"province"
				]
			},
			"AddressRequest": {
				"type": "object",
				"properties": {
					"label": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"postal_code": {
						"type": "string"
					},
					"is_default": {
						"type": "boolean"
					}
				},
				"required": [
					"address",
					"city",
					"label",
					"name",
					"phone",
					"postal_code",
					"province"
				]
			}
		}
	},
	"paths": {
...
		"/addresses": {
			"get": {
				"summary": "List the caller's addresses, default first",
				"tags": [
					"addresses"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Address"
									}
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Add an address to the caller's account",
				"tags": [
					"addresses"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"requestBody": {
					"description": "Address",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressRequest",
								"summary": "body",
								"description": "Address"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Address"
								}
							}
						}
					},
					"400": {
						"description": "Invalid body",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				}
			}
		}
	}
}